+++ com.onesignal.unity.android/Runtime/AndroidUserManager.cs
@@ -51,6 +51,11 @@ namespace OneSignalSDK.Android.User {
             set => _user.Call("setLanguage", value);
         }
 
+        public Dictionary<string, string> GetTags() {
+            AndroidJavaObject obj = _user.Call<AndroidJavaObject>("getTags");
+            return obj.MapToDictionary();
+        }
+
         public void AddTag(string key, string value)
             => _user.Call("addTag", key, value);