+++ com.onesignal.unity.ios/Runtime/iOSUserManager.cs
@@ -44,6 +44,7 @@ namespace OneSignalSDK.iOS.User {
         [DllImport("__Internal")] private static extern void _oneSignalUserRemoveEmail(string email);
         [DllImport("__Internal")] private static extern void _oneSignalUserAddSms(string smsNumber);
         [DllImport("__Internal")] private static extern void _oneSignalUserRemoveSms(string smsNumber);
+        [DllImport("__Internal")] private static extern string _oneSignalUserGetTags();
         [DllImport("__Internal")] private static extern void _oneSignalUserAddTag(string key, string value);
         [DllImport("__Internal")] private static extern void _oneSignalUserAddTags(string tagsJson);
         [DllImport("__Internal")] private static extern void _oneSignalUserRemoveTag(string key);
@@ -61,6 +62,13 @@ namespace OneSignalSDK.iOS.User {
 
         public string Language {
             set => _oneSignalUserSetLanguage(value);
+        }
+
+        public Dictionary<string, string> GetTags() {
+            Dictionary<string, object> raw = (Dictionary<string, object>)Json.Deserialize(_oneSignalUserGetTags());
+            Dictionary<string, string> dict = new Dictionary<string, string>();
+            foreach (KeyValuePair<string, object> kvp in raw) dict.Add(kvp.Key, kvp.Value.ToString());
+            return dict;
         }
 
         public void AddTag(string key, string value)